Version 4
Not sure what changed, but I think its better:

To do:

Need to import taruns photos of the top 100 meals. ( its in jsonl format to need to convert to typescript and add it to my file)

Need to fix the meal card positioning

Also need to add the text blob to the front, and have it flip over on tap and show ingredients and macros on the back.

Fix the ingredients and meals button on cart, it so far only works when clicked on in the center

Work on the filter section